Move date utils into directory
