+++ src/Components/Dashboard.js
@@ -4,8 +4,8 @@ import { useCurrentUserContext } from "../Contexts/currentUserContext";
 import { useCurrentOrganizationContext } from "../Contexts/currentOrganizationContext";
 import { getAllUsers } from "../Services/Organizations/OrganizationUsersService";
 import { getAllGrants } from "../Services/Organizations/GrantsService";
-import isRecent from "../Helpers/isRecent";
-import isSoon from "../Helpers/isSoon";
+import isRecent from "../Helpers/date/isRecent";
+import isSoon from "../Helpers/date/isSoon";
 import Button from "./design/Button/Button";
 import CurrentOrganizationLink from "./Helpers/CurrentOrganizationLink";
 import Spinner from "./Helpers/Spinner";
